requirements.txt
allure-python-commons
allure-pytest
flake8
pytest~=8.0.0
selenium